--- templates/contacts.html
@@ -1,37 +1,0 @@
-{{ define "contacts"}}
-{{ template "header"}}
-<h3>По всем вопросам обращаться</h3>
- <p>
-   <span>Баканов роман</span>
-   <ul>
-     <li>+7 (999) 999 99 99</li>
-     <li>@RealLoh</li>
-     <li>ya.loh@mail.ru</li>
-     <li>тут ссылка</li>
-   </ul>
- </p>
- <p>
-   <span>Гендина Нина</span>
-   <ul>
-     <li>+7 (985) 271 44 07</li>
-     <li>@Ninel'</li>
-     <li>gendina.nina@mail.ru</li>
-     <li>тут ссылка</li>
-   </ul>
- </p>
- <p>
-   Офис
- </p>
- <div class="">
-   тут карта офиса кста
- </div>
-
-
-
- <style media="screen">
-   .act3 {
-     color: #fff !important;
-     border-bottom-color: #fff !important;
-   }
- </style>
-{{ end }}
