--- templates/footer.html
@@ -1,15 +1,0 @@
-{{ define "footer"}}
-      <br>
-      <br>
-      <footer class="mt-auto text-white-50 foot">
-        <p>Production from
-          <a href="ee" class="text-white">Nina</a>, by
-          <a  href="ee" class="text-white">Roma</a>.
-        </p>
-      </footer>
-    </div>
-    <script src="../static/js/jquery.min.js"></script>
-    <script src="../static/js/site.js"></script>
-  </body>
-</html>
-{{ end }}
