--- templates/form1.html
@@ -1,28 +1,0 @@
-
-  <div class="user-box">
-    <input type="email" name="mail" required maxlength="50">
-    <label>Email</label>
-  </div>
-  <div class="user-box">
-    <input type="text" name="name" required maxlength="40">
-    <label>Name</label>
-  </div>
-  <div class="user-box">
-    <input type="text" name="surname" required maxlength="40">
-    <label>Surname</label>
-  </div>
-  <div class="user-box">
-    <input type="password" name="password" required maxlength="40" minlength="6">
-    <label>Password</label>
-  </div>
-  <div class="user-box">
-    <input type="password" name="repeat_password" required maxlength="40" minlength="6">
-    <label>Repeat password</label>
-  </div>
-  <a href="#" onclick='change_forms();'>
-    <span></span>
-    <span></span>
-    <span></span>
-    <span></span>
-    Далее
-  </a>
