--- templates/form2.html
@@ -1,29 +1,0 @@
-
-  <div class="user-box">
-    <input type="date" name="birthday" required="" max="2003-01-01" min="1921-01-01">
-  </div>
-  <div class="user-box">
-    <input type="text" required maxlength="40">
-    <label>Job</label>
-  </div>
-  <div class="user-box">
-    <input type="text" required maxlength="40">
-    <label>University</label>
-  </div>
-  <div class="radio">
-    <input type="radio" name="sex" value="m" class="sex1" autofocus checked><span class="label">M</span>
-    <input type="radio" name="sex" value="w" class="sex"><span class="label">W</span>
-    <input type="radio" name="sex" value="o" class="sex"><span class="label">Другое</span>
-  </div>
-  <br>
-  <div class="user-box">
-    <input type="text" required>
-    <label>Чтот еще</label>
-  </div>
-  <a href="#" onclick='change_forms_second();'>
-    <span></span>
-    <span></span>
-    <span></span>
-    <span></span>
-    Далее
-  </a>
