--- templates/my_page.html
@@ -1,44 +1,0 @@
-{{ define "my_page"}}
-{{ template "header" .Session}}
-<h1>{{ .User.Name }}</h1>
-<div class="main">
-  <div class="photo_profile">
-    <img id="img" src="https://sun9-64.userapi.com/impg/RJYL18z_li7U_6R1hef_VMEnehYfBE1GKkvz3Q/sn3chH1oCME.jpg?size=607x1024&quality=95&sign=2aca13ec6e1d01336d289aef9ede1319&type=album" alt="">
-  </div>
-  <div class="info">
-    <div class="basic_info">
-      <h3>25</h3>
-      <h3>основная</h3>
-      <h3>ииииинфа</h3>
-    </div>
-    <div class="additional_info">
-      <h3>25</h3>
-      <h3>доп</h3>
-      <h3>ииииинфа</h3>
-    </div>
-    <div class="tags">
-      <div class="tag">
-        <span>крутая</span>
-      </div>
-      <div class="tag">
-        <span>класная</span>
-      </div>
-    </div>
-  </div>
-  <div class="map">
-    <img id="img_map" src="https://sun9-29.userapi.com/impg/y5npyYstu_z0JDhcN3Dn3VWh8-yLPTmbdhK2kw/4ETlZByvuSE.jpg?size=1876x1156&quality=96&sign=008fc31281482cd39a1bb6c31f31c904&type=album" id="" alt="">
-    <div class="">
-      <a href="/log_out">Выход</a>
-      <a href="/favorite">Фавориты</a>
-      <a href="#">Настройки</a>
-    </div>
-  </div>
-</div>
-<style media="screen">
-  .act1 {
-    color: #fff !important;
-    border-bottom-color: #fff !important;
-  }
-</style>
-{{ template "footer"}}
-{{ end }}
