--- templates/user_card.html
@@ -1,25 +1,0 @@
-{{ define "user_card"}}
-<div class="card">
-    <div class="photo" href="/profile/{{ .Id }}">
-      <a href="/profile/{{ .Id }}" id="link_card">
-      <img src="https://sosedi-ng-rb.s3.eu-west-2.amazonaws.com/{{ .Photo }}?AWSAccessKeyId=KIASJYGCYOCKR2SATBN" alt="" class="img">
-    </a>
-    </div>
-    <div class="info">
-      <p class="user_id" hidden>{{ .Id }}</p>
-      <span class="name">{{ .Name }}, </span>
-      <span class="age">25</span><br>
-      <span class="price">{{ .PriceMin }} - {{ .PriceMax }} ₽</span><br>
-      <span class="likes_number">20 ♥</span>
-    </div>
-    <div class="icon_wraper" hidden>
-      <img src="../static/img/heart.svg" alt="" class="icon">
-    </div>
-</div>
-<style media="screen">
-  #link_card {
-    text-decoration: none !important;
-    color: inherit !important;
-  }
-</style>
-{{ end }}
