+++ views/account/authorization.html
@@ -1,0 +1,97 @@
+{{ define "authorization"}}
+{{ template "header_log"}}
+<div id="errors">
+{{ . }}
+</div>
+<div class="login-box">
+  <h2>SOSEDI</h2>
+  <form id="myform" method="post" action="/log_in">
+    <div class="user-box">
+      <input ttype="email"  name="email" required>
+      <label>Email</label>
+    </div>
+    <div class="user-box">
+      <input type="password" name="password" required maxlength="40" minlength="6">
+      <label>Password</label>
+    </div>
+    <a href="#" onclick="send();">
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      Log in
+    </a>
+  </form>
+</div>
+<div id="exist">
+  Еще нет аккаунта? <a href="/registration">Зарегестрируйтесь!</a>
+</div>
+<style media="screen">
+  html {
+    height: 100%;
+  }
+  body {
+    margin:0;
+    padding:0;
+    font-family: sans-serif;
+    background: linear-gradient(rgb(34,37, 41), #003457);
+  }
+
+  #errors {
+    color: white;
+    background: #732124;
+    text-align: center;
+    font-size: 1.3em;
+  }
+
+  #exist {
+    position: absolute;
+    top: 73%;
+    left: 42%;
+    color: rgba(0,0,0,.5);
+  }
+  #exist > a {
+    text-decoration: none;
+    color: #005f9e;
+  }
+</style>
+<script type="text/javascript">
+  function send() {
+        var pattern = /^[a-z0-9][a-z0-9\._-]*[a-z0-9]*@([a-z0-9]+([a-z0-9-]*[a-z0-9]+)*\.)+[a-z]+/i;
+        var mail = $('input[name=email]');
+            if(mail.val() != ''){
+                // Проверяем, если введенный email соответствует регулярному выражению
+                if(mail.val().search(pattern) == 0){
+                    // Убираем сообщение об ошибке
+                    $('#errors').text('');
+                }else{
+                    //Выводим сообщение об ошибке
+                    $('#errors').text('Введите корректный Email, пожалуйста');
+                    return
+                }
+            }else{
+                $('#errors').text('Введите Ваш email, пожалуйста');
+                return
+            }
+        //================ Проверка длины пароля ==================
+        var password = $('input[name=password]');
+            if(password.val() != ''){
+                //Если длина введенного пароля меньше шести символов, то выводим сообщение об ошибке
+                if(password.val().length < 6){
+                    //Выводим сообщение об ошибке
+                    $('#errors').text('Пароль не может быть короче 6 символов');
+                    return
+                }else{
+                    // Убираем сообщение об ошибке
+                    $('#errors').text('');
+                }
+            }else{
+                $('#errors').text('Введите свой пароль, пожалуйста');
+                return
+            }
+        document.getElementById('myform').submit(); return false;
+  }
+</script>
+</body>
+</html>
+{{ end }}
