+++ views/favorite.html
@@ -1,0 +1,22 @@
+{{ define "favorite"}}
+{{ template "header" .Session}}
+<p class="lead">
+  <a href="#" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Взаимно</a>
+  <a href="#" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Не заимно</a>
+  <a href="#" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Все</a>
+</p>
+
+{{ range $user := .Users }}
+  {{ template "long_card" $user }}
+
+<br><br><br><br>
+<!-- {{ template "long_card" }} -->
+{{ end }}
+<style media="screen">
+  .act2 {
+    color: #fff !important;
+    border-bottom-color: #fff !important;
+  }
+</style>
+{{ template "footer" }}
+{{ end }}
