+++ views/shared/header.html
@@ -1,0 +1,26 @@
+{{ define "header"}}
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+<head>
+  <meta charset="utf-8">
+  <title>Sosedi</title>
+  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.1/dist/css/bootstrap.min.css">
+  <link rel="stylesheet" href="https://getbootstrap.com/docs/5.1/examples/cover/cover.css">
+  <link rel="shortcut icon" href="../static/img/gay2.png" type="image/x-icon">
+  <link rel="stylesheet" href="../static/css/main.css">
+</head>
+<body class="text-center text-white bg-dark">
+  <div class="d-flex w-100 h-100 p-7 mx-auto flex-column">
+    <header class="mb-auto">
+      <div>
+        <a href="/" class="a_title"><h3 class="float-md-start mb-0" id="title">Sosedi</h3></a>
+        <nav class="nav nav-masthead justify-content-center float-md-end">
+          {{ if eq . ""}}
+          {{ template "../shared/header_not_authorized" }}
+          {{ else }}
+          {{ template "../shared/header_authorized" .}}
+          {{ end }}
+        </nav>
+      </div>
+    </header>
+{{ end }}
