+++ views/shared/header_authorized.html
@@ -1,0 +1,5 @@
+{{ define "header_authorized"}}
+<a class="nav-link act1" href="/my_page">My Page</a>
+<a class="nav-link act2" href="/favorite">Favorite</a>
+<a class="nav-link act3" href="/contacts">Contact</a>
+{{ end }}
