+++ views/shared/header_log.html
@@ -1,0 +1,14 @@
+{{ define "header_log"}}
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+<head>
+  <meta charset="utf-8">
+  <title>Sosedi</title>
+  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.1/dist/css/bootstrap.min.css">
+  <link rel="stylesheet" href="https://getbootstrap.com/docs/5.1/examples/cover/cover.css">
+  <link rel="shortcut icon" href="../static/img/gay2.png" type="image/x-icon">
+  <link rel="stylesheet" href="../static/css/log.css">
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+</head>
+<body>
+{{ end }}
