+++ views/shared/header_not_authorized.html
@@ -1,0 +1,4 @@
+{{ define "header_not_authorized"}}
+<a class="nav-link act1" href="/registration">Зарегестрироваться</a>
+<a class="nav-link act2" href="/authorization">Войти</a>
+{{ end }}
